add average comments per post to items analytics

+++ api/typeDefs/growth.js
@@ -3,7 +3,7 @@ import { gql } from 'graphql-tag'
 export default gql`
   type NameValue {
     name: String!
-    value: Int!
+    value: Float!
   }
 
   extend type Query {
